Configuration: Improved handling of sections and values, fixing some bugs.
The property tree model caused issues because sections and keys on identical
paths occupy the same storage in the tree, but may refer to different things
with different context etc. Due to this, methods such as erase(Path) did not
work well, deleting both keys and whole sections, causing zombie ptree nodes
to appear etc.

Hence we introduce a new value type to the property tree: TreeItem, which
contains either or both of the section information or value information.

Also removed PathNotFoundException.